Allow replies to comments and deleting comments only replaces text

+++ src/definitions/schemas/mongoose/comment.ts
@@ -44,12 +44,20 @@ commentSchema.index(
   { unique: true }
 );
 
-commentSchema.pre(['deleteMany', 'findOneAndDelete'], async function() {
+commentSchema.pre(['deleteMany'], async function() {
   const results = await Comment.find(this.getQuery(), '_id');
   const _ids = results.map(comment => comment._id);
+  if (_ids.length === 0) return;
+
   /* Delete all votes on Comment.  */
   await Vote.deleteMany({
     votedModel: 'Comment',
+    postId: { $in: _ids }
+  });
+
+  /* Delete all replies to Comment.  */
+  await Comment.deleteMany({
+    commentedModel: 'Comment',
     postId: { $in: _ids }
   });
 });
